Remove signInWithApple method and add example
