+++ Sources/GoTrue/Internal/Helpers.swift
@@ -45,34 +45,3 @@ private func base64URLDecode(_ value: String) -> Data? {
   }
   return Data(base64Encoded: base64, options: .ignoreUnknownCharacters)
 }
-
-func randomString(length: Int = 32) -> String {
-  precondition(length > 0)
-  var randomBytes = [UInt8](repeating: 0, count: length)
-  let errorCode = SecRandomCopyBytes(kSecRandomDefault, randomBytes.count, &randomBytes)
-  if errorCode != errSecSuccess {
-    fatalError(
-      "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
-    )
-  }
-
-  let charset: [Character] =
-    Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-
-  let nonce = randomBytes.map { byte in
-    // Pick a random character from the set, wrapping around if needed.
-    charset[Int(byte) % charset.count]
-  }
-
-  return String(nonce)
-}
-
-func sha256(_ input: String) -> String {
-  let inputData = Data(input.utf8)
-  let hashedData = SHA256.hash(data: inputData)
-  let hashString = hashedData.compactMap {
-    String(format: "%02x", $0)
-  }.joined()
-
-  return hashString
-}
